+++ association.go
@@ -1,0 +1,6 @@
+package trig
+
+type Association struct {
+	TriggeringFileName string
+	DependentFileNames []string
+}
