+++ trig_test.go
@@ -1,0 +1,9 @@
+package trig
+
+import (
+	"testing"
+)
+
+func TestSomething(t *testing.T) {
+	t.Fatal("test not implemented")
+}
